@@ -55,17 +55,17 @@
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^6.0.4",
-    "@angular/compiler-cli": "^6.0.5",
-    "@ngtools/webpack": "^6.0.8",
+    "@angular/cli": "^6.1.0-beta.1",
+    "@angular/compiler-cli": "^6.1.0-beta.1",
+    "@ngtools/webpack": "^6.1.0-beta.1",
     "@types/common-tags": "^1.4.0",
     "@types/d3": "^5.0.0",
     "@types/debug": "0.0.30",
     "@types/electron-config": "^0.2.1",
     "@types/electron-devtools-installer": "^2.0.3",
     "@types/electron-window-state": "^2.0.31",
-    "@types/lodash": "^4.14.109",
-    "@types/node": "^10.3.3",
+    "@types/lodash": "^4.14.110",
+    "@types/node": "^10.3.4",
     "@types/request-promise-native": "^1.0.15",
     "copy-webpack-plugin": "^4.5.1",
     "copyfiles": "^2.0.0",
@@ -92,7 +92,7 @@
     "to-string-loader": "^1.1.5",
     "tslib": "^1.9.2",
     "tslint": "^5.10.0",
-    "typescript": "^2.7.2",
+    "typescript": "^2.8.4",
     "webpack": "^4.12.0",
     "webpack-cli": "^3.0.8"
   }
