@@ -1,13 +1,26 @@
 import { Injectable } from '@angular/core'
-import { Router, CanActivate, CanActivateChild } from '@angular/router'
+import {
+  ActivatedRoute,
+  ActivatedRouteSnapshot,
+  CanActivate,
+  CanActivateChild,
+  Router,
+  RouterStateSnapshot
+} from '@angular/router'
 
 @Injectable()
-export class ApiAccessGuard implements CanActivate {
-  constructor(private readonly router: Router) {}
+export class ApiAccessGuard implements CanActivate, CanActivateChild {
+  constructor(private readonly router: Router, private route: ActivatedRoute) {}
 
-  canActivate() {
-    // this.router.navigate(['login']);
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
+    if (!!0) this.router.navigate(['login'])
     console.log('CanActivate')
+    console.log(route, state, this.route)
+    return true
+  }
+
+  canActivateChild() {
+    console.log('CanActivateChild')
     return true
   }
 }
